Reduces overall LoRA loading times significantly when using a different base model with `--lora-base`, from 32s to 24s in my test case.

It also seems to improve the general performance of `ggml_cpy` significantly, about twice as fast, but overall this is an insignificant fraction of the eval time, so it isn't really noticeable.

I tried to cover all the paths in `ggml_cpy`, but there are a lot of them and only a few are hit in llama.cpp, so I have not tested every single one.

Perplexity (bs=512):
```
MASTER: perf_total_per_op_us[             CPY] =  309.170 ms
PR:     perf_total_per_op_us[             CPY] =  132.353 ms
```

LoRA (quantize):
```
MASTER: perf_total_per_op_us[             CPY] =  45.780 ms
PR:     perf_total_per_op_us[             CPY] =   5.255 ms
```